Update README.md
Number Guessing Game is a simple game is which is designed using core Java in this we used the java concepts like classes objects abstaraction interfaces java swings and Java AWT etc. The Game functionalities the computer will generate random number which we have to guess. we have 10 lifes, for every incorrect guess we lose one life for every correct guess we get 3 more lifes.We have higest score and for every time there we will get corresponding message etc.